# Use @swc/jest to speed up tests