--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -17,7 +17,6 @@ commands:
       - node_install
       - node/install-packages:
           include-branch-in-cache-key: true
-          cache-path: ~/project/node_modules
           override-ci-command: npm install
       - checkout_core
       - run: