--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -17,7 +17,7 @@ commands:
       - node_install
       - node/install-packages:
           include-branch-in-cache-key: true
-          override-ci-command: npm install
+          cache-path: ~/project/node_modules
       - checkout_core
       - run:
           name: Set GB locale