--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -11,13 +11,24 @@ commands:
           name: Clone bichard7-next-core
           command: git clone --depth 1 https://github.com/ministryofjustice/bichard7-next-core.git ~/bichard7-next-core
 
+  install_npm_packages:
+    steps:
+        - restore_cache:
+            name: Load node_modules from the cache if they haven't changed
+            keys:
+              - v1-npm-deps-{{ .Branch  }}-{{ checksum "package-lock.json" }}
+              - v1-npm-deps-
+        - run:
+            name: Install npm dependencies
+            command: npm i
+        - save_cache:
+            key: v1-npm-deps-{{ .Branch  }}-{{ checksum "package-lock.json" }}
+            paths: node_modules
   setup:
     steps:
       - checkout
       - node_install
-      - node/install-packages:
-          include-branch-in-cache-key: true
-          cache-path: ~/project/node_modules
+      - install_npm_packages
       - checkout_core
       - run:
           name: Set GB locale
@@ -179,8 +190,7 @@ jobs:
     steps:
       - checkout
       - node_install
-      - node/install-packages:
-          include-branch-in-cache-key: true
+      - install_npm_packages
       - run:
           name: Check the code for linting errors
           command: npm run lint
@@ -192,8 +202,7 @@ jobs:
     steps:
       - checkout
       - node_install
-      - node/install-packages:
-          include-branch-in-cache-key: true
+      - install_npm_packages
       - run:
           name: Run code-based unit tests
           command: npm run test:unit
@@ -246,8 +255,7 @@ jobs:
             - checkout_core
             - install_ms_edge
       - node_install
-      - node/install-packages:
-          include-branch-in-cache-key: true
+      - install_npm_packages
       - run:
           name: Install Cypress
           command: ./.circleci/scripts/install-cypress.sh 2>&1