--- a/.swcrc
+++ b/.swcrc
@@ -1,0 +1,17 @@
+{
+  "jsc": {
+    "parser": {
+      "syntax": "typescript",
+      "tsx": false,
+      "decorators": true
+    },
+    "transform": {
+      "legacyDecorator": true,
+      "decoratorMetadata": true
+    },
+    "target": "esnext"
+  },
+  "module": {
+    "type": "commonjs"
+  }
+}