- [x] Unit tests
- [x] Integration tests
- [x] Tweak the npm run tests config
- [x] Cache node modules properly in CircleCI properly

On Circle CI:

Main: 
- Total time of job for unit tests = 1min 40s
- Unit tests = 32s
- Total time of job for integration tests = 4m 29s
- Integration tests = 1m

This branch (with cached npm modules):
- Total time of job for unit tests = 31s
- Unit tests = 4s
- Total time of job for integration tests = 1m 55s
- Integration tests = 39s